requirements.txt
numpy
pandas
scipy
giotto-tda
plotly
dash
discord.py
websockets
aiohttp
questdb
uvloop
